Refactored plugin menus to use new ui class AdvancedMenuSelector
