+++ JKMP.Core/Configuration/Attributes/PropertyCreators/Implementations/TextFieldCreator.cs
@@ -16,7 +16,7 @@ namespace JKMP.Core.Configuration.Attributes.PropertyCreators.Implementations
             typeof(string)
         };
 
-        public override IMenuItem CreateField(object config, string fieldName, PropertyInfo propertyInfo, TextFieldAttribute attribute, MenuSelector menu, List<IDrawable> drawables)
+        public override IMenuItem CreateField(object config, string fieldName, PropertyInfo propertyInfo, TextFieldAttribute attribute, List<IDrawable> drawables)
         {
             var result = new TextInputField(fieldName, (string)propertyInfo.GetValue(config), attribute.MaxLength, JKContentManager.Font.MenuFont)
             {
