+++ JKMP.Core/Configuration/UI/IConfigMenu.cs
@@ -1,4 +1,6 @@
 using System.Collections.Generic;
+using BehaviorTree;
+using JKMP.Core.UI;
 using JumpKing.PauseMenu;
 using JumpKing.PauseMenu.BT;
 using JumpKing.Util;
@@ -7,7 +9,7 @@ namespace JKMP.Core.Configuration.UI
 {
     public interface IConfigMenu
     {
-        public MenuSelector CreateMenu(GuiFormat format, string name, MenuSelector parent, List<IDrawable> drawables);
+        public IBTnode CreateMenu(GuiFormat format, string name, AdvancedMenuSelector parent, List<IDrawable> drawables);
     }
     
     public interface IConfigMenu<T> : IConfigMenu where T : class, new()
